floating number move animation

+++ src/HonkTrooper/HonkTrooper/Constructs/HealthyConstruct.cs
@@ -4,7 +4,7 @@
     {
         public double Health { get; set; }
 
-        public double HitPoint { get; set; } = 5;
+        public int HitPoint { get; set; } = 5;
 
         public bool IsDead => Health <= 0;
     }
